refactor: migrate route_export_test e2e
